+++ src/animation.js
@@ -626,7 +626,7 @@ function stopAnimationAll(done){
 }
 // end of stop
 
-// Pause animation
+// Pause and resume animation
 function pauseAnimation(id){
     var an = $.animation.elements[id];
 
@@ -645,14 +645,19 @@ function pauseAnimation(id){
 function pauseAnimationAll(filter){
     $.each($.animation.elements, function(k, v){
         if (filter) {
-            if (matches.call(v.element, filter)) pauseAnimation(k);
+            if (typeof filter === "string") {
+                if (matches.call(v.element, filter)) pauseAnimation(k);
+            } else if (filter.length) {
+                $.each(filter, function(){
+                    if (v.element === this) pauseAnimation(k);
+                });
+            }
         } else {
             pauseAnimation(k);
         }
     });
 }
-// end og pause
-
+// end of pause
 
 function resumeAnimation(id){
     var an = $.animation.elements[id];
@@ -672,7 +677,13 @@ function resumeAnimation(id){
 function resumeAnimationAll(filter){
     $.each($.animation.elements, function(k, v){
         if (filter) {
-            if (matches.call(v.element, filter)) resumeAnimation(k);
+            if (typeof filter === "string") {
+                if (matches.call(v.element, filter)) resumeAnimation(k);
+            } else if (filter.length) {
+                $.each(filter, function(){
+                    if (v.element === this) resumeAnimation(k);
+                });
+            }
         } else {
             resumeAnimation(k);
         }